restructure bootloader
s1_mbr goes in the MBR of a disk, loads the first active
partition's 1st sector to 0x7C00, and jumps to it. s1_mbr does
not provide any other bootstrapping or functionality

s1_vbr is the first sector of a FAT32-formatted disk. It
assumes that s1_mbr, or possibly another bootloader, has done
nothing to the system and it'll need to do everything for itself;
this behaviour allows it to be used as an MBR as well. s1_vbr
will load a file with a specific name from the FAT32 filesystem.

s1_mbr is done. work on s1_vbr is underway.